Event optimization

+++ todo-list/script.js
@@ -10,6 +10,7 @@ const todo = document.querySelector("#todo");
 todo.addEventListener("submit", (e) => {
   e.preventDefault();
   const todoTitle = todo.querySelector("input").value;
+
   if (todoTitle === "") {
     return;
   }
@@ -19,8 +20,6 @@ todo.addEventListener("submit", (e) => {
   todoList.push({ title: todoTitle, id: Date.now() });
   render();
 });
-
-render();
 
 function deleteTodoFromList(id) {
   todoList = todoList.filter((item) => item.id !== id);
@@ -46,8 +45,10 @@ function render() {
     const noItem = document.createElement("h1");
     noItem.innerText = "Ooops! List is empty";
     todoContainer.appendChild(noItem);
+    localStorage.setItem("todos", JSON.stringify(todoList));
     return;
   }
+
   todoList.forEach(({ title, id }) => {
     const todoItem = document.createElement("div");
     todoItem.setAttribute("class", "todo-item");
@@ -55,6 +56,10 @@ function render() {
 
     const todoTitle = document.createElement("h4");
     todoTitle.innerText = title;
+
+    const allButtons = document.createElement("div");
+    allButtons.setAttribute("class", "todo-buttons");
+
     const editTodo = document.createElement("button");
     editTodo.setAttribute("title", "Edit");
     editTodo.setAttribute("id", "edit-btn");
@@ -66,33 +71,49 @@ function render() {
     deleteTodo.innerHTML = `<i class="fa fa-trash" aria-hidden="true"></i>`;
 
     todoItem.appendChild(todoTitle);
-    todoItem.appendChild(editTodo);
-    todoItem.appendChild(deleteTodo);
+    allButtons.appendChild(deleteTodo);
+    allButtons.appendChild(editTodo);
+    todoItem.appendChild(allButtons);
 
-    // Edit todo
-    editTodo.addEventListener("click", (e) => {
-      todoTitle.contentEditable = true;
-      todoTitle.focus();
-      const saveTodo = document.createElement("button");
-      saveTodo.setAttribute("id", "save-btn");
-      todoTitle.setAttribute("class", "save-input");
-      saveTodo.innerHTML = `<i class="fa fa-bookmark" aria-hidden="true"></i>`;
-      todoItem.appendChild(saveTodo);
+    // Event delegation
+    todoItem.addEventListener("click", (e) => {
+      if (
+        e.target.id == "delete-btn" ||
+        e.target.parentElement?.id == "delete-btn"
+      ) {
+        // Delete todo
+        deleteTodoFromList(id);
+      } else if (
+        e.target.id == "edit-btn" ||
+        e.target.parentElement?.id == "edit-btn"
+      ) {
+        // edit todo
+        todoTitle.contentEditable = true;
+        todoTitle.focus();
+        const saveTodo = document.createElement("button");
+        saveTodo.setAttribute("id", "save-btn");
+        todoTitle.setAttribute("class", "save-input");
+        saveTodo.innerHTML = `<i class="fa fa-bookmark" aria-hidden="true"></i>`;
+        allButtons.appendChild(saveTodo);
 
-      saveTodo.addEventListener("click", (e) => {
-        todoTitle.contentEditable = false;
-        todoItem.removeChild(saveTodo);
-        todoTitle.removeAttribute("class", "save-input");
-        const updatedTodoTitle = todoTitle.innerText;
-        updateTodoFromList(id, updatedTodoTitle);
-      });
-    });
-
-    // Delete todo
-    deleteTodo.addEventListener("click", () => {
-      deleteTodoFromList(id);
+        saveTodo.addEventListener("click", (e) => {
+          todoTitle.contentEditable = false;
+          allButtons.removeChild(saveTodo);
+          todoTitle.removeAttribute("class", "save-input");
+          const updatedTodoTitle = todoTitle.innerText;
+          updateTodoFromList(id, updatedTodoTitle);
+        });
+      }
     });
 
     todoContainer.appendChild(todoItem);
   });
+
+  localStorage.setItem("todos", JSON.stringify(todoList));
 }
+
+// After load DOM call render function
+(() => {
+  todoList = JSON.parse(localStorage.getItem("todos")) || todoList;
+  render();
+})();
